removed redirect from AddOrder
