+++ src/App.js
@@ -51,7 +51,7 @@ function App() {
             <Route path="/scan/:id" component={ScanOrder} />
             <Route path="/orders/:id" component={EditOrder} />
             <AdminRoute exact path="/users/add" component={AddUser} />
-            <FedRoute exact path="/add" component={AddOrder} />
+            <Route exact path="/add" component={AddOrder} />
             {/* Refresh should be removed prior to production */}
             <Route exact path="/refresh" component={Refresh} />{" "}
           </UserContext.Provider>
